+++ CHANGELOG.md
@@ -14,7 +14,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Switch to Circle CI Rust 1.33.0 image.
 
 ### Fixed
-* Fix erroneous doc comments (#19).
+* Fix erroneous doc comments (PR #24).
+* Unimplement `Clone`, `Send`, and `Sync` for `RenderDoc` (PR #29).
 
 ## [0.4.0] - 2018-09-16
 ### Added